Move http library implementations to core directory.
Use lib directory only to provide API to plugin.
