+++ test/response.ts
@@ -1,5 +1,5 @@
 import assert = require('assert');
-import Response = require("../lib/response");
+import Response = require("../core/http/response");
 
 describe('Response', function () {
   describe('#constructor()', function () {
